dashboard/requirements.txt
streamlit==1.15.1
pandas==1.4.3
matplotlib==3.5.3
seaborn==0.11.2
cassandra-driver==3.25.0
pymongo==4.2.0
plotly==5.10.0